chore: use regex to assert component state (#1885)
In flow-components 24.8 a new property has been added to some field components
causing test failures because assertions are based on the textual representation
of the component and the introduced attribute is not expected.
This change updates the assertion to use regular expression to check only the
relevant part of the component representation.